Generate Shiny app

This function generates the Shiny app. It is the main function of the package. The app can either be launched directly, or saved to a folder. If saved, that folder will contain all the code and data needed to run the app. The app can then be launched by sourcing the app.R file in the folder and running launch_app().

Usage

generate_shiny(
  dataset,
  dataset_name,
  eff_size_type_label = NA,
  models = get_model_tibble,
  save_to_folder = NA,
  launch_app = is.na(save_to_folder),
  ...
)

Arguments

dataset

Dataframe as returned from import_data()

dataset_name

Name of the dataset

eff_size_type_label

The label to be used to describe the effect size. If NA, effect type code from dataset is used.

models

The models to be included in the app. Can either be a function to call, a tibble, or the path to a file. If you want to change the default models, have a look at the vignette and/or the get_model_tibble() documentation. Passing a file (e.g. "my_models.R") is particularly helpful if you include helper functions and save the app. If so, this must assign the tibble to a global variable called models_to_run (i.e. using <<-).

save_to_folder

If specified, the code and data for the app will be saved to this folder. Defaults to NA, which means that nothing will be saved. If the folder exists, the user will be asked to confirm overwriting it - unless the script is run in non-interactive mode, in which case the folder will be overwritten without asking.

launch_app

Should the app be launched? Defaults to TRUE if it is not saved (i.e. save_to_folder is NA), FALSE otherwise.

...

Arguments passed on to create_about

date

Date of the last update. Defaults to the current date.

citation

Citation requested by the author. Defaults to an empty string.

osf_link

Link to the OSF (or similar) project where data and materials are available. Defaults to an empty string.

contact

Contact information, typically email. Defaults to an empty string.